Move namespace to ...Configuration.Internal
