+++ src/Microsoft.EntityFrameworkCore.Tools.Core/Microsoft.EntityFrameworkCore.Tools.Core.csproj
@@ -87,23 +87,23 @@
     <Compile Include="Properties\ToolsCoreStrings.Designer.cs">
       <DependentUpon>ToolsCoreStrings.resx</DependentUpon>
     </Compile>
+    <Compile Include="Scaffolding\Configuration\Internal\AttributeConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\CustomConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\EntityConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\FluentApiConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\IAttributeConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\IFluentApiConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\IndexConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\KeyFluentApiConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\ModelConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\NavigationPropertyConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\NavigationPropertyInitializerConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\OptionsBuilderConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\PropertyConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\RelationshipConfiguration.cs" />
+    <Compile Include="Scaffolding\Configuration\Internal\SequenceConfiguration.cs" />
     <Compile Include="Scaffolding\Internal\CodeWriter.cs" />
     <Compile Include="Scaffolding\Internal\ConfigurationFactory.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\AttributeConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\CustomConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\EntityConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\FluentApiConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\IAttributeConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\IFluentApiConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\IndexConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\KeyFluentApiConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\ModelConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\NavigationPropertyConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\NavigationPropertyInitializerConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\OptionsBuilderConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\PropertyConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\RelationshipConfiguration.cs" />
-    <Compile Include="Scaffolding\Internal\Configuration\SequenceConfiguration.cs" />
     <Compile Include="Scaffolding\Internal\DbContextWriter.cs" />
     <Compile Include="Scaffolding\Internal\EntityTypeWriter.cs" />
     <Compile Include="Scaffolding\Internal\FileSystemFileService.cs" />
