+++ src/Microsoft.EntityFrameworkCore.Tools.Core/Scaffolding/Configuration/Internal/NavigationPropertyConfiguration.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Utilities;
 
-namespace Microsoft.EntityFrameworkCore.Scaffolding.Internal.Configuration
+namespace Microsoft.EntityFrameworkCore.Scaffolding.Configuration.Internal
 {
     public class NavigationPropertyConfiguration
     {
