+++ src/Microsoft.EntityFrameworkCore.Tools.Core/Scaffolding/Configuration/Internal/NavigationPropertyInitializerConfiguration.cs
@@ -4,7 +4,7 @@
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Utilities;
 
-namespace Microsoft.EntityFrameworkCore.Scaffolding.Internal.Configuration
+namespace Microsoft.EntityFrameworkCore.Scaffolding.Configuration.Internal
 {
     public class NavigationPropertyInitializerConfiguration
     {
