+++ src/Microsoft.EntityFrameworkCore.Tools.Core/Scaffolding/Internal/StringBuilderCodeWriter.cs
@@ -5,7 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
-using Microsoft.EntityFrameworkCore.Scaffolding.Internal.Configuration;
+using Microsoft.EntityFrameworkCore.Scaffolding.Configuration.Internal;
 using Microsoft.EntityFrameworkCore.Utilities;
 
 namespace Microsoft.EntityFrameworkCore.Scaffolding.Internal
